+++ CHANGELOG.md
@@ -107,6 +107,7 @@ git-cliff 0e4e3abba94636c210fde5e7d02089298589f26d..HEAD --output CHANGELOG.md
 ## 1.11.2 - 2023-02-15
 ### Changed
 - Refactored: symbols redesign #116
+- Refactored: report default text #283
 ### Fixed
 - Fixed: random sorting on persons #237
 - Fixed: linking persons to report #262